src/AgentBin.Python/requirements.txt
a2a-sdk[http-server,sqlite]>=0.3.0
uvicorn>=0.32.0
fastapi>=0.115.0
